+++ src/main/java/btw/lowercase/skyboxify/screen/SkyLayerListScreen.java
@@ -42,9 +42,9 @@ public class SkyLayerListScreen extends DebugScreen {
 	public SkyLayerListScreen(final Screen parent, final Skybox skybox) {
 		super(Component.literal(
 				//? >=1.21.11 {
-				/*skybox.getWorldKey().location().toString()
+				/*skybox.getDimension().location().toString()
 				 *///?} else {
-				skybox.getWorldKey().location().toString()
+				skybox.getDimension().location().toString()
 				//?}
 		), parent);
 		this.skybox = skybox;